DocuQuery_agent/requirements.txt
phidata
python-dotenv
openai
qdrant-client
bs4
sqlalchemy
duckduckgo-search
packaging
pypdf